Some more work towards double-precision support. Most of these changes are to preserve coordinate's data type as it passes through the filter.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@19049 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/operators/MetricThreshold/avtMetricThresholdFilter.h
@@ -87,7 +87,7 @@ class avtMetricThresholdFilter : public avtPluginDataTreeIterator
 
     vtkDataSet          *ExecuteOnScalarData(vtkDataSet *, vtkDataArray *);
     vtkPoints           *CreateRectilinearPoints(vtkRectilinearGrid *);
-    inline bool          PassesTest(int type, float value);
+    inline bool          PassesTest(int type, double value);
 };
